refactor: modularize dataset loaders
